fix: matching constructors to wii u
i noticed constructor are different. still nothing changed :(

+++ Minecraft.World/RareBiomeLayer.h
@@ -5,7 +5,7 @@
 class RareBiomeLayer : public Layer
 {
 public:
-	RareBiomeLayer(int64_t seed, shared_ptr<Layer> parent);
+	RareBiomeLayer(int32_t seed, shared_ptr<Layer> parent);
 	virtual ~RareBiomeLayer() {}
 	virtual intArray getArea(int xo, int yo, int w, int h) override;
 };
